+++ src/canisters/converter.ts
@@ -3,7 +3,7 @@ import { sha224 } from "js-sha256";
 import { Buffer } from "buffer";
 // @ts-ignore (no type definitions for crc are available)
 import crc from "crc";
-import { SUB_ACCOUNT_BYTE_LENGTH } from "./constants";
+import { SUB_ACCOUNT_BYTE_LENGTH } from "../constants/constants";
 import { AccountIdentifier, SubAccount } from "../types/common";
 
 export const uint8ArrayToBigInt = (array: Uint8Array): bigint => {
